+++ src/app/group/team/[id]/page.tsx
@@ -1,17 +1,15 @@
 'use client';
 
 import { useTheme } from '@/hooks/useTheme';
-import { Avatar, AvatarFallback, AvatarImage } from '@/shadcn/ui/avatar';
-import { Button } from '@/shadcn/ui/button';
-import { Card, CardContent, CardHeader } from '@/shadcn/ui/card';
-import { Separator } from '@/shadcn/ui/separator';
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/shadcn/ui/tooltip';
-import { Progress } from '@/shadcn/ui/progress';
+import { Card, CardContent } from '@/shadcn/ui/card';
 import { spacing } from '@/styles/design-system';
-import { CheckIcon, Edit, Plus, Target, Trophy } from 'lucide-react';
 import { useParams } from 'next/navigation';
 import { useState, useCallback, useMemo } from 'react';
 import DateNavigation from '@/components/common/DateNavigation';
+import TeamCard from '@/components/group/TeamCard';
+import GroundRules from '@/components/group/GroundRules';
+import TeamLeaderboard from '@/components/group/TeamLeaderboard';
+import TodayGoals from '@/components/group/TodayGoals';
 import { getKSTDateString, getKSTDateStringDaysAgo } from '@/utils/timezone';
 
 export default function TeamDetailPage() {
@@ -19,28 +17,6 @@ export default function TeamDetailPage() {
   const params = useParams();
   const teamId = params.id;
 
-  const [todayGoals] = useState([
-    {
-      id: 1,
-      title: 'Complete React refactoring task',
-      achieved: ['John Doe', 'Sarah Wilson'],
-      notAchieved: ['Mike Johnson', 'Jane Smith', 'Tom Brown', 'Alice Kim']
-    },
-    {
-      id: 2,
-      title: 'Review team code submissions', 
-      achieved: ['John Doe', 'Mike Johnson', 'Jane Smith'],
-      notAchieved: ['Sarah Wilson', 'Tom Brown', 'Alice Kim']
-    },
-    {
-      id: 3,
-      title: 'Update project documentation',
-      achieved: ['Sarah Wilson', 'Tom Brown'],
-      notAchieved: ['John Doe', 'Mike Johnson', 'Jane Smith', 'Alice Kim']
-    }
-  ]);
-  const [isEditingGoal, setIsEditingGoal] = useState(false);
-  const [newGoal, setNewGoal] = useState('');
   const [selectedPeriod, setSelectedPeriod] = useState<'daily' | 'weekly'>('daily');
   const [selectedDate, setSelectedDate] = useState(getKSTDateString());
 
@@ -74,65 +50,29 @@ export default function TeamDetailPage() {
       'Code reviews within 24 hours',
       'No meetings on Fridays',
       'Use proper commit messages'
+    ],
+    todayGoals: [
+      {
+        id: 1,
+        title: 'Complete React refactoring task',
+        achieved: ['John Doe', 'Sarah Wilson'],
+        notAchieved: ['Mike Johnson', 'Jane Smith', 'Tom Brown', 'Alice Kim']
+      },
+      {
+        id: 2,
+        title: 'Review team code submissions', 
+        achieved: ['John Doe', 'Mike Johnson', 'Jane Smith'],
+        notAchieved: ['Sarah Wilson', 'Tom Brown', 'Alice Kim']
+      },
+      {
+        id: 3,
+        title: 'Update project documentation',
+        achieved: ['Sarah Wilson', 'Tom Brown'],
+        notAchieved: ['John Doe', 'Mike Johnson', 'Jane Smith', 'Alice Kim']
+      }
     ]
   };
 
-  const getAvatarInitials = (name: string) => {
-    return name.split(' ').map(n => n[0]).join('');
-  };
-
-  const renderAvatarGroup = (names: string[], isAchieved: boolean) => {
-    const maxVisible = 3;
-    const totalAvatars = names.length;
-    const displayedNames = names.slice(0, maxVisible);
-    const remainingCount = Math.max(totalAvatars - maxVisible, 0);
-
-    return (
-      <div className="flex items-center">
-        {displayedNames.map((name, index) => (
-          <div key={index} className={`-ml-2 hover:z-10 relative first:ml-0`} style={{ zIndex: displayedNames.length - index }}>
-            <Avatar className="w-6 h-6 ring-1 ring-gray-200 dark:ring-gray-700">
-              <AvatarImage src="" />
-              <AvatarFallback className={`text-[8px] font-semibold ${getThemeClass('component')} ${getThemeTextColor('primary')}`}>
-                {getAvatarInitials(name)}
-              </AvatarFallback>
-            </Avatar>
-          </div>
-        ))}
-        {remainingCount > 0 && (
-          <TooltipProvider>
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <Avatar className="w-6 h-6 -ml-2 hover:z-10 relative ring-1 ring-gray-200 dark:ring-gray-700 cursor-pointer">
-                  <AvatarFallback className={`text-[8px] font-bold ${getThemeClass('component')} ${getThemeTextColor('primary')}`}>
-                    +{remainingCount}
-                  </AvatarFallback>
-                </Avatar>
-              </TooltipTrigger>
-              <TooltipContent side="top" className="max-w-xs">
-                <div className="space-y-1">
-                  <div className="text-xs font-medium mb-2">
-                    {isAchieved ? 'Also Achieved:' : 'Also Not Achieved:'}
-                  </div>
-                  {names.slice(maxVisible).map((name, i) => (
-                    <div key={i} className="text-xs">
-                      {name}
-                    </div>
-                  ))}
-                </div>
-              </TooltipContent>
-            </Tooltip>
-          </TooltipProvider>
-        )}
-      </div>
-    );
-  };
-
-  const handleGoalSave = () => {
-    setNewGoal('');
-    setIsEditingGoal(false);
-    // TODO: API로 새 목표 추가 로직 구현
-  };
 
   const handlePreviousDate = useCallback(() => {
     const currentIndex = availableDates.indexOf(selectedDate);
@@ -193,71 +133,14 @@ export default function TeamDetailPage() {
   return (
     <div className="h-full grid grid-cols-5 gap-6 p-6" style={{ gridTemplateRows: 'auto auto 1fr' }}>
       {/* 좌측 최상단 - 그룹 이름 및 팀장 정보 (3/5) */}
-      <Card className={`${getCommonCardClass()} col-span-3 row-span-1`}>
-        <CardContent className={spacing.inner.normal}>
-          <div className="space-y-4">
-            <div className="flex items-center gap-3">
-              <div className={`text-2xl font-bold ${getThemeTextColor('primary')}`}>
-                {teamData.name}
-              </div>
-              <div className="flex gap-2">
-                <span className={`px-2 py-1 text-xs font-medium rounded-full ${getThemeClass('componentSecondary')} ${getThemeTextColor('secondary')}`}>
-                  #Frontend
-                </span>
-                <span className={`px-2 py-1 text-xs font-medium rounded-full ${getThemeClass('componentSecondary')} ${getThemeTextColor('secondary')}`}>
-                  #React
-                </span>
-              </div>
-            </div>
-            
-            <div className="flex items-center space-x-3">
-              <Avatar className="w-12 h-12">
-                <AvatarImage src="" />
-                <AvatarFallback className={`text-sm font-semibold ${getThemeClass('componentSecondary')} ${getThemeTextColor('primary')}`}>
-                  {teamData.leader.avatar}
-                </AvatarFallback>
-              </Avatar>
-              <div>
-                <p className={`font-medium ${getThemeTextColor('primary')}`}>
-                  {teamData.leader.name}
-                </p>
-                <p className={`text-sm ${getThemeTextColor('secondary')}`}>
-                  Team Leader
-                </p>
-              </div>
-            </div>
-            
-            <div className={`mt-4 p-3 rounded-lg ${getThemeClass('componentSecondary')}`}>
-              <p className={`text-sm ${getThemeTextColor('secondary')}`}>
-                {teamData.description}
-              </p>
-            </div>
-          </div>
-        </CardContent>
-      </Card>
+      <TeamCard 
+        teamName={teamData.name}
+        description={teamData.description}
+        leader={teamData.leader}
+      />
 
       {/* 우측 상단 - 그라운드 룰 (2/5) */}
-      <Card className={`${getCommonCardClass()} col-span-2 row-span-2`}>
-        <CardHeader>
-          <div className={`text-lg font-bold ${getThemeTextColor('primary')}`}>
-            Ground Rules
-          </div>
-        </CardHeader>
-        <CardContent className={spacing.inner.normal}>
-          <div className="space-y-4">
-            {teamData.groundRules.map((rule, index) => (
-              <div key={index} className={`flex items-start gap-3 p-3 rounded-md ${getThemeClass('componentSecondary')} hover:${getThemeClass('component')} transition-colors`}>
-                <div className={`w-6 h-6 rounded-full flex items-center justify-center text-xs font-bold flex-shrink-0 ${getThemeClass('component')} ${getThemeTextColor('secondary')}`}>
-                  {index + 1}
-                </div>
-                <span className={`text-sm ${getThemeTextColor('primary')} leading-relaxed`}>
-                  {rule}
-                </span>
-              </div>
-            ))}
-          </div>
-        </CardContent>
-      </Card>
+      <GroundRules rules={teamData.groundRules} />
 
       {/* Navigation Controls - 별도 컴포넌트 */}
       <Card className={`${getCommonCardClass()} col-span-3 row-span-1 flex items-center`}>
@@ -303,163 +186,10 @@ export default function TeamDetailPage() {
       </Card>
 
       {/* 좌측 하단 - 리더보드 (3/5) */}
-      <Card className={`${getCommonCardClass()} col-span-3 row-span-1`}>
-        <CardHeader>
-          <div className={`text-lg font-bold ${getThemeTextColor('primary')}`}>
-            Leaderboard
-          </div>
-        </CardHeader>
-        <CardContent className={spacing.inner.normal}>
-          <div className="grid grid-cols-3 gap-4">
-            {teamData.leaderboard.map((member) => (
-              <div key={member.rank} className={`p-4 rounded-md ${getThemeClass('componentSecondary')} text-center relative`}>
-                
-                {/* 아바타 */}
-                <div className="relative w-fit mx-auto mb-3">
-                  <Avatar className="w-16 h-16 ring-1 ring-gray-200 dark:ring-gray-700">
-                    <AvatarImage src="" />
-                    <AvatarFallback className={`text-lg font-semibold ${getThemeClass('component')} ${getThemeTextColor('primary')}`}>
-                      {member.name.split(' ').map(n => n[0]).join('')}
-                    </AvatarFallback>
-                  </Avatar>
-                  <span className={`absolute -right-0 -bottom-0 size-4 rounded-full ${
-                    member.rank <= 3 
-                      ? 'bg-green-500 dark:bg-green-400' 
-                      : 'bg-gray-400 dark:bg-gray-500'
-                  }`}>
-                  </span>
-                </div>
-                
-                {/* 정보 */}
-                <div>
-                  <p className={`text-sm font-medium truncate ${getThemeTextColor('primary')} mb-1`}>
-                    {member.name}
-                  </p>
-                  <p className={`text-xs ${getThemeTextColor('secondary')} mb-1`}>
-                    {member.hours}h worked
-                  </p>
-                  <p className={`text-lg font-bold ${getThemeTextColor('primary')}`}>
-                    {member.score}
-                  </p>
-                </div>
-              </div>
-            ))}
-          </div>
-        </CardContent>
-      </Card>
+      <TeamLeaderboard members={teamData.leaderboard} />
 
       {/* 우측 하단 - 오늘의 목표 설정 (2/5) */}
-      <Card className={`${getCommonCardClass()} col-span-2 row-span-1`}>
-        <CardHeader className="flex flex-row items-center justify-between">
-          <div className={`text-lg font-bold ${getThemeTextColor('primary')}`}>
-            Today's Goal
-          </div>
-          <Button
-            size="sm"
-            variant="ghost"
-            onClick={() => setIsEditingGoal(true)}
-            className={`h-8 w-8 p-0 ${getThemeTextColor('secondary')} hover:${getThemeClass('componentSecondary')}`}
-          >
-            <Plus className="h-4 w-4" />
-          </Button>
-        </CardHeader>
-        <CardContent className={spacing.inner.normal}>
-          <Separator className="mb-4" />
-          <div className="space-y-4">
-            {isEditingGoal ? (
-              <div className="space-y-3">
-                <input
-                  type="text"
-                  value={newGoal}
-                  onChange={(e) => setNewGoal(e.target.value)}
-                  className={`w-full px-3 py-2 rounded-md border ${getThemeClass('border')} ${getThemeClass('component')} ${getThemeTextColor('primary')} focus:outline-none focus:ring-2 focus:ring-[#3F72AF]`}
-                  placeholder="Enter a new goal..."
-                  autoFocus
-                />
-                <div className="flex gap-2">
-                  <Button
-                    onClick={handleGoalSave}
-                    disabled={!newGoal.trim()}
-                    className={`${getThemeClass('componentSecondary')} ${getThemeTextColor('primary')} hover:${getThemeClass('component')}`}
-                  >
-                    Add Goal
-                  </Button>
-                  <Button
-                    variant="outline"
-                    onClick={() => {
-                      setIsEditingGoal(false);
-                      setNewGoal('');
-                    }}
-                    className={`${getThemeClass('component')} ${getThemeClass('border')} ${getThemeTextColor('secondary')}`}
-                  >
-                    Cancel
-                  </Button>
-                </div>
-              </div>
-            ) : (
-              <div className="space-y-4">
-                {todayGoals.map((goal, index) => (
-                  <div key={goal.id} className={`p-3 rounded-lg ${getThemeClass('componentSecondary')}`}>
-                    <div className="flex items-center justify-between">
-                      <div className="flex items-center gap-2">
-                        <div className={`w-5 h-5 rounded-full flex items-center justify-center text-xs font-bold ${getThemeClass('component')} ${getThemeTextColor('secondary')}`}>
-                          {index + 1}
-                        </div>
-                        <div className={`text-sm font-bold ${getThemeTextColor('primary')}`}>
-                          {goal.title}
-                        </div>
-                      </div>
-                      <div className="flex items-center gap-2">
-                        <div className="flex items-center gap-2">
-                          <div className="relative w-24">
-                            <Progress 
-                              value={(goal.achieved.length / (goal.achieved.length + goal.notAchieved.length)) * 100} 
-                              className="h-2 bg-gray-200"
-                            />
-                            <div className="absolute inset-0 h-2 rounded-full bg-green-500" 
-                              style={{ width: `${(goal.achieved.length / (goal.achieved.length + goal.notAchieved.length)) * 100}%` }}
-                            />
-                          </div>
-                          <span className={`text-xs font-bold ${getThemeTextColor('primary')}`}>
-                            {Math.round((goal.achieved.length / (goal.achieved.length + goal.notAchieved.length)) * 100)}%
-                          </span>
-                        </div>
-                        <span className={`text-xs font-medium ${getThemeTextColor('secondary')}`}>
-                          {goal.achieved.length}/{goal.achieved.length + goal.notAchieved.length}
-                        </span>
-                      </div>
-                    </div>
-                    
-                    <div className="flex gap-4 mt-3">
-                      {/* 달성한 사람들 */}
-                      <div className="flex-1">
-                        <div className="flex items-center gap-2 mb-2">
-                          <span className={`text-xs font-medium ${getThemeTextColor('secondary')}`}>
-                            Achieved ({goal.achieved.length})
-                          </span>
-                        </div>
-                        {renderAvatarGroup(goal.achieved, true)}
-                      </div>
-
-                      <Separator orientation="vertical" className="h-16" />
-
-                      {/* 달성하지 못한 사람들 */}
-                      <div className="flex-1">
-                        <div className="flex items-center gap-2 mb-2">
-                          <span className={`text-xs font-medium ${getThemeTextColor('secondary')}`}>
-                            Not Achieved ({goal.notAchieved.length})
-                          </span>
-                        </div>
-                        {renderAvatarGroup(goal.notAchieved, false)}
-                      </div>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-        </CardContent>
-      </Card>
+      <TodayGoals goals={teamData.todayGoals} />
     </div>
   );
 }
